added svcontrol function

+++ tests/__init__.py
@@ -3,9 +3,10 @@ import os
 import signal
 import shutil
 import textwrap
-from subprocess import Popen, PIPE
 import time
-from supervise_web.io import svstat
+import datetime
+from subprocess import Popen, PIPE
+from supervise_web.io import svstat, svcontrol
 
 
 class BaseTestCase(TestCase):
@@ -20,9 +21,11 @@ class BaseTestCase(TestCase):
     def tearDown(self):
         # shut down all test daemons and stop their supervise processes
         for num, supervise_proc in self.supervise_procs.items():
-            daemon_pid = svstat(self._daemon_dir(num))['pid']
+            daemon_pid = svstat(self._daemon_dir(num))['daemon_pid']
             supervise_proc.terminate()
-            os.kill(daemon_pid, signal.SIGTERM)
+            # only kill daemon if still alive (and not killed by some test)
+            if os.path.exists('/proc/%d' % daemon_pid):
+                os.kill(daemon_pid, signal.SIGTERM)
 
     def _daemon_dir(self, num):
         return os.path.join(self.test_dir, 'test_daemon_%d' % num)
@@ -45,7 +48,7 @@ class BaseTestCase(TestCase):
         supervise_proc = Popen(['supervise', self._daemon_dir(num)], stdin=PIPE, stderr=PIPE)
         max_wait = 2  # wait for at most 2 seconds for the supervise process to start
         while max_wait > 0:
-            if svstat(self._daemon_dir(num))['supervise_alive']:
+            if svstat(self._daemon_dir(num))['alive']:
                 break
             max_wait -= 0.1
             time.sleep(0.1)
@@ -60,7 +63,3 @@ class BaseTestCase(TestCase):
         with open(self._daemon_run_file(num), 'w') as f:
             f.write(daemon_code)
         os.chmod(self._daemon_run_file(num), 0700)
-
-    def test_test_setup(self):
-        for i in range(1):
-            self._create_test_daemon(i)
